+++ planetarium-story-unity/Assets/Scripts/TableSheet/ShopCharacterSheet.cs
@@ -6,12 +6,12 @@
         {
             public override int Key => TotalCount;
             public int TotalCount { get; private set; }
-            public int Cost { get; private set; }
+            public long Cost { get; private set; }
 
             public override void Set(string[] fields)
             {
                 TotalCount = ParseInt(fields[0]);
-                Cost = ParseInt(fields[1]);
+                Cost = ParseLong(fields[1]);
             }
         }
     }
