+++ planetarium-story-unity/Assets/Scripts/TableSheet/ShopSpaceSheet.cs
@@ -6,16 +6,15 @@
         {
             public override int Key => Step;
             public int Step { get; private set; }
-            public int Cost { get; private set; }
+            public long Cost { get; private set; }
             public float BonusSpeed { get; private set; }
 
             public override void Set(string[] fields)
             {
                 Step = ParseInt(fields[0]);
-                Cost = ParseInt(fields[1]);
+                Cost = ParseLong(fields[1]);
                 BonusSpeed = ParseFloat(fields[2]);
             }
         }
-        
     }
 }
